Store OpenGL texture staging areas as `uint32_t` vectors

+++ src/gui/render/opengl_renderer.h
@@ -17,6 +17,7 @@
 
 #include "dosbox_config.h"
 #include "gui/render/render.h"
+#include "misc/video.h"
 #include "utils/rect.h"
 
 // Glad must be included before SDL
@@ -52,7 +53,7 @@ public:
 
 	std::string GetCurrentShaderDescriptorString() override;
 
-	void StartFrame(uint8_t*& pixels_out, int& pitch_out) override;
+	void StartFrame(uint32_t*& pixels_out, int& pitch_out) override;
 	void EndFrame() override;
 
 	void PrepareFrame() override;
@@ -134,10 +135,15 @@ private:
 
 	// The current framebuffer we render the emulated video output into
 	// (contains the "work-in-progress" next frame).
-	std::vector<uint8_t> curr_framebuf = {};
+	//
+	// The framebuffers contain 32-bit pixel data stored as a sequence of
+	// four packed 8-bit values in BGRX byte order (that's in memory order,
+	// so byte N is B, byte N+1 is G, byte N+2 is R).
+	//
+	std::vector<uint32_t> curr_framebuf = {};
 
 	// Contains the last fully rendered frame, waiting to be presented.
-	std::vector<uint8_t> last_framebuf = {};
+	std::vector<uint32_t> last_framebuf = {};
 
 	// True if the last framebuffer has been updated since the last present
 	bool last_framebuf_dirty = false;
